90_TeamProject/01_PaperReview_240115/team_2/SWEngineer/requirements.txt
langchain
langchain-chroma
langchain-openai
langchain-teddynote
langchain-community
langchain-experimental

python-dotenv
pysqlite3-binary
streamlit
pymupdf
python-dotenv
chromadb
pymongo
pydantic
pyyaml